client/requirements.txt
PyInquirer==1.0.3
python-dotenv==0.13.0
requests==2.23.0
